requirements.txt
streamlit>=1.32.0
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.38.0
ffmpeg-python>=0.2.0
numpy>=1.24.0
scikit-learn>=1.3.0
plotly>=5.18.0